pyproject.toml
[project]
name = "cyberguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "e2b-code-interpreter>=0.0.10",
    "llama-index-llms-fireworks>=0.2.1",
    "llama-index>=0.11.16",
    "python-dotenv>=1.0.1",
    "streamlit>=1.39.0",
]
